+++ backend/canisters/user/impl/src/openchat_bot.rs
@@ -3,7 +3,7 @@ use crate::updates::c2c_send_messages::{HandleMessageArgs, handle_message_impl};
 use chat_events::{MessageContentInternal, TextContentInternal};
 use constants::{OPENCHAT_BOT_USER_ID, OPENCHAT_BOT_USERNAME};
 use types::{ChannelId, CommunityId, EventWrapper, Message, User, UserId, UserType};
-use user_canister::{C2CReplyContext, PhoneNumberConfirmed, StorageUpgraded, UserSuspended};
+use user_canister::C2CReplyContext;
 use user_core::openchat_bot;
 
 pub(crate) fn send_community_deleted_message(deleted_by: UserId, name: String, public: bool, state: &mut RuntimeState) {
@@ -47,23 +47,6 @@ pub(crate) fn send_removed_from_group_or_community_message(
     let text =
         openchat_bot::removed_from_group_or_community_text(is_group, removed_by, &group_or_community_name, public, blocked);
     send_text_message(text, Vec::new(), false, state);
-}
-
-pub(crate) fn send_phone_number_confirmed_bot_message(event: &PhoneNumberConfirmed, state: &mut RuntimeState) {
-    send_text_message(openchat_bot::phone_number_confirmed_text(event), Vec::new(), false, state);
-}
-
-pub(crate) fn send_storage_ugraded_bot_message(event: &StorageUpgraded, state: &mut RuntimeState) {
-    send_text_message(openchat_bot::storage_upgraded_text(event), Vec::new(), false, state);
-}
-
-pub(crate) fn send_referred_user_joined_message(user_id: UserId, username: String, state: &mut RuntimeState) {
-    let text = openchat_bot::referred_user_joined_text(user_id);
-    send_text_message(text, vec![User { user_id, username }], false, state);
-}
-
-pub(crate) fn send_user_suspended_message(event: &UserSuspended, state: &mut RuntimeState) {
-    send_text_message(openchat_bot::user_suspended_text(event), Vec::new(), false, state);
 }
 
 pub(crate) fn send_message(